// src/app/model/costs-model.ts
export interface CostsModel {
  countryName: string;
  vehicle: string;
  fuel_type: string;
  weight_cost: number;
  distance_cost: number;
  worktime_cost: number;
  heightDimension_cost: number;
  widthDimension_cost: number;
  lengthDimension_cost: number;
  average_consumption_cost: number;
  countryNew: boolean;
  statusChange: boolean;
}