// src/app/model/tariff-plan-model.ts
export interface TariffPlanModel {

  weight_plan: boolean;
  distance_plan: boolean;
  worktime_plan: boolean;
  fuelconsumption_plan: boolean;
  tolls_plan: boolean;
  dimension_plan: boolean;
  antt_plan: boolean;
  fragile_plan: boolean;
  persishable_plan: boolean;
  reshipping_plan: boolean;
}